1、添加三方组件的功能 2、解决Cpu向导重名的问题 3、解决新建的板件不能成功创建工程的问题 4、信息未填完整时，提示用户
git-svn-id: https://xiangmuzuserver/svn/%E7%A1%AC%E4%BB%B6%E7%BB%84%E5%BC%80%E5%8F%91%E5%BA%93/platform_soft/djyos/DIDE/EclipsePluginSource@42802 98a3a3af-8837-1c49-9e9e-13921c22655a

+++ org.eclipse.cdt-CDT_9_3_1/core/org.eclipse.cdt.ui/src/org/eclipse/cdt/ui/wizards/board/ReadBoardXml.java
@@ -58,10 +58,10 @@ public class ReadBoardXml {
 						cpuOn.setCpuName(content);
 						break;
 					case "mianClk":
-						cpuOn.setMianClk(Integer.parseInt(content));
+						cpuOn.setMianClk(Float.parseFloat(content));
 						break;
 					case "rtcClk":
-						cpuOn.setRtcClk(Integer.parseInt(content));
+						cpuOn.setRtcClk(Float.parseFloat(content));
 						break;
 					case "chip":
 						NodeList chipList = cNode.getChildNodes();
